requirements.txt
streamlit==1.32.2
transformers==4.38.2
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
spacy==3.7.4
detoxify==0.5.2
torch
torchvision
pandas
sentencepiece
python-dotenv
google-genai
openai
anthropic